refactor(linter): Put all tasks back in top-level `linter.py`
After discussion, we would like to keep all task definitions directly in `.py` files directly in `tasks/`.
All helper functions and modules will be under `libs/` instead.
