+++ tasks/libs/linter/__init__.py
@@ -1,0 +1,1 @@
+"""Helper functions for linting tasks"""
