--- tasks/linter/gitlab/__init__.py
@@ -1,29 +1,0 @@
-"""Linting-related tasks for files related to Gitlab CI"""
-
-# Make tasks & helper methods accessible as they are used in unit test mocks
-from . import helpers, tasks
-from .tasks import (
-    gitlab_change_paths,
-    gitlab_ci,
-    gitlab_ci_jobs_codeowners,
-    gitlab_ci_jobs_needs_rules,
-    gitlab_ci_jobs_owners,
-    gitlab_ci_shellcheck,
-    job_change_path,
-    list_parameters,
-    ssm_parameters,
-)
-
-__all__ = [
-    "helpers",
-    "tasks",
-    "gitlab_change_paths",
-    "gitlab_ci",
-    "gitlab_ci_jobs_codeowners",
-    "gitlab_ci_jobs_needs_rules",
-    "gitlab_ci_jobs_owners",
-    "gitlab_ci_shellcheck",
-    "job_change_path",
-    "list_parameters",
-    "ssm_parameters",
-]
